+++ src/components/icon/__spec__.js
@@ -190,6 +190,46 @@ describe('Icon', () => {
       svg = span.children[0];
       expect(svg.getAttribute('class')).toEqual("ui-icon__svg ui-icon__svg--chevron");
     });
+
+    it('calls the render information icon method', () => {
+      instance = TestUtils.renderIntoDocument(<Icon type='information' />);
+      instance.renderIcon;
+      span = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'span')[1];
+      svg = span.children[0];
+      expect(svg.getAttribute('class')).toEqual("ui-icon__svg ui-icon__svg--information");
+    });
+
+    it('calls the render sync icon method', () => {
+      instance = TestUtils.renderIntoDocument(<Icon type='sync' />);
+      instance.renderIcon;
+      span = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'span')[1];
+      svg = span.children[0];
+      expect(svg.getAttribute('class')).toEqual("ui-icon__svg ui-icon__svg--sync");
+    });
+
+    it('calls the render progress icon method', () => {
+      instance = TestUtils.renderIntoDocument(<Icon type='progress' />);
+      instance.renderIcon;
+      span = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'span')[1];
+      svg = span.children[0];
+      expect(svg.getAttribute('class')).toEqual("ui-icon__svg ui-icon__svg--progress");
+    });
+
+    it('calls the render submitted icon method', () => {
+      instance = TestUtils.renderIntoDocument(<Icon type='submitted' />);
+      instance.renderIcon;
+      span = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'span')[1];
+      svg = span.children[0];
+      expect(svg.getAttribute('class')).toEqual("ui-icon__svg ui-icon__svg--submitted");
+    });
+
+    it('calls the render completed icon method', () => {
+      instance = TestUtils.renderIntoDocument(<Icon type='completed' />);
+      instance.renderIcon;
+      span = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'span')[1];
+      svg = span.children[0];
+      expect(svg.getAttribute('class')).toEqual("ui-icon__svg ui-icon__svg--completed");
+    });
   });
 
   describe('success', () => {